WIP: Add my expansion material
This is just a local WIP. DO NOT PUSH
This is a first draft of an expansion material that expands linear in
time. This a first step towards a reaction material.

Maybe it should be replaced by using the general multiplicative
split material. But this could have other disadvantages depending on
how I e.g. get reaction rate to my poro

DO NOT PUSH!!!