WIP: Use my exp mat with a scalar
This is a temporary solution which makes no sense long term. I change
my existing exp material (which wasn't great to start) to base the
expansion on a scalar. There is already an existing material that can
do this which I should use but it doesn't work with the element I am
using. So this a temp hack to test things but once I have time I should
use MAT_MultiplicativeSplitDefgradElastHyper +
MAT_InelasticDefgradLinScalarIso - but this only works for one scalar
so can't do reaction and thermal expansion