HACK: Change n scalars to 3 for nitsche contact
This is a hack it increases the number of scalars in ssi and pssi
nitsche contact to 3. This means usinbg one scalar is no longer
possible. This BREAKS EXISTING USES!

Can't find a great way to have num scalar variable as it ise used in
const expressions, templates etc. that are fixed at compile time.

I could make it work with leaving kss etc the same size as for one but
then would need a new assembly function that uses num scalar to put it
at right spot (but would be a lot of  effort)

HACK THAT BREAKS STUFF

Change which scalar transported nitsche pssi to the third
(thats the +2 part)